Modified the rmTrash.

+++ rmTrash.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 
-users=("seg" "ipr")
+# Check if at least one user name is provided
+if [ "$#" -lt 1 ]; then
+    echo "Usage: $0 user1 [user2 ...]"
+    exit 1
+fi
 
-for user in "${users[@]}"; do
+# Loop through each user name provided as argument
+for user in "$@"; do
     sudo rm -rf /home/"$user"/.local/share/Trash/files
-    sudo mkdir /home/"$user"/.local/share/Trash/files
-    
+    sudo mkdir -p /home/"$user"/.local/share/Trash/files
 done
